feat: move /bin/template to /templates
