+++ lib/create.js
@@ -53,7 +53,7 @@ module.exports.createProject = function (project_path, package_name, project_nam
     }
 
     // copy template/* to project_path directory ( recursive )
-    fs.cpSync(path.join(ROOT, 'bin/template'), project_path, { recursive: true });
+    fs.cpSync(path.join(ROOT, 'templates'), project_path, { recursive: true });
 
     // TODO: stop copying cordova-browser logic into platforms/browser/
     // recreate our node_modules structure in the new project
@@ -74,18 +74,18 @@ module.exports.createProject = function (project_path, package_name, project_nam
 
     // copy cordova js file to platform_www
     fs.cpSync(
-        path.join(ROOT, 'bin/template/www/cordova.js'),
+        path.join(ROOT, 'templates/www/cordova.js'),
         path.join(platform_www, 'cordova.js')
     );
 
     // copy favicon file to platform_www
     fs.cpSync(
-        path.join(ROOT, 'bin/template/www/favicon.ico'),
+        path.join(ROOT, 'templates/www/favicon.ico'),
         path.join(platform_www, 'favicon.ico')
     );
 
     // load manifest to write name/shortname
-    const manifest = require(path.join(ROOT, 'bin/template/www', 'manifest.json'));
+    const manifest = require(path.join(ROOT, 'templates/www', 'manifest.json'));
     manifest.name = project_name;
     manifest.short_name = project_name;
     // copy manifest file to platform_www
