+++ templates/cordova/defaults.xml
