+++ templates/cordova/version
