+++ templates/www/favicon.ico
