+++ templates/www/manifest.json
